add default visualizer for audio
