+++ app/javascript/styles/mastodon/components.scss
@@ -775,14 +775,41 @@
       padding: 8px;
     }
 
-    &__preview {
+    &__preview,
+    &__visualizer {
       position: absolute;
       width: 100%;
       height: 100%;
-      border-radius: 6px;
       z-index: -1;
       top: 0;
+    }
+
+    &__preview {
+      border-radius: 6px;
       inset-inline-start: 0;
+    }
+
+    &__visualizer {
+      padding: 16px;
+      box-sizing: border-box;
+
+      .audio-player__visualizer {
+        margin: 0 auto;
+        display: block;
+        height: 100%;
+      }
+
+      .icon-music {
+        position: absolute;
+        top: 50%;
+        inset-inline-start: 50%;
+        transform: translate(-50%, -50%);
+        opacity: 0.75;
+        color: var(--player-foreground-color);
+        filter: var(--overlay-icon-shadow);
+        width: 48px;
+        height: 48px;
+      }
     }
 
     &__thumbnail {
